Add first version of the pygeppetto API
This first version had been generated from the last
geppettoModel.ecore version from the 'developpement' branch. It should
be noted that a copy of the 'geppettoModel.ecore' had been provided in
order to ensure that the version that have been used for the
generation is still in the same state.