Add a demo file and some test data
The test data are related to the developement URI of the Geppetto
metamodel. It seems that the URI of the Geppetto metamodel master
version is different. Without any extra mapping in a resource
set/global registry, the opening of the current xmi-data tests will
failed with the master URI (if the URIs of the master version matches
the ones from the master Geppetto metamodel).