Setup project for packaging and tests
This commit introduced many things related to packaging and tests.

Regardings the tests, I choose pytest which is a well known and
practical test framework. As complement, the coverage plugin had been
enabled. This plugin keep tracks of the lines executed by the
tests. To run all the tests, tox is used as it integrate very well
with travis and it creates virtualenv to run the tests.

For packaging purposes, setuptools is used.