Add registration of the master/develop URI in the global_registry
The URIs of metamodels need to be registered for actually reading a
new model. This registration need to be performed by hand when a new
ResourceSet is built. This commit introduces an auto-registration in
the global_registry of the URI for the developement branch and the
master one (as the metamodel URI are different regarding the branch
they are embedded in). The fact to use the global_registry avoid the
need to register the URI/metamodel in each ResourceSet.